+++ src/main/java/uk/gov/hmcts/pdda/business/AbstractControllerBean.java
@@ -10,7 +10,9 @@ import uk.gov.hmcts.pdda.business.entities.xhbconfigprop.XhbConfigPropRepository
 import uk.gov.hmcts.pdda.business.entities.xhbcourt.XhbCourtRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbcppformatting.XhbCppFormattingRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbcpplist.XhbCppListRepository;
+import uk.gov.hmcts.pdda.business.entities.xhbcppstaginginbound.XhbCppStagingInboundRepository;
 import uk.gov.hmcts.pdda.business.entities.xhbformatting.XhbFormattingRepository;
+import uk.gov.hmcts.pdda.business.entities.xhbpddamessage.XhbPddaMessageRepository;
 
 public class AbstractControllerBean {
 
@@ -24,18 +26,23 @@ public class AbstractControllerBean {
     private XhbCppFormattingRepository xhbCppFormattingRepository;
     private XhbCppListRepository xhbCppListRepository;
     private XhbFormattingRepository xhbFormattingRepository;
+    private XhbPddaMessageRepository xhbPddaMessageRepository;
+    private XhbCppStagingInboundRepository xhbCppStagingInboundRepository;
 
     // For unit tests.
     protected AbstractControllerBean(EntityManager entityManager,
         XhbClobRepository xhbClobRepository, XhbBlobRepository xhbBlobRepository,
         XhbCourtRepository xhbCourtRepository, XhbConfigPropRepository xhbConfigPropRepository,
-        XhbCppFormattingRepository xhbCppFormattingRepository) {
+        XhbCppFormattingRepository xhbCppFormattingRepository, XhbPddaMessageRepository xhbPddaMessageRepository,
+            XhbCppStagingInboundRepository xhbCppStagingInboundRepository) {
         this(entityManager);
         this.xhbClobRepository = xhbClobRepository;
         this.xhbBlobRepository = xhbBlobRepository;
         this.xhbCourtRepository = xhbCourtRepository;
         this.xhbConfigPropRepository = xhbConfigPropRepository;
         this.xhbCppFormattingRepository = xhbCppFormattingRepository;
+        this.xhbPddaMessageRepository = xhbPddaMessageRepository;
+        this.xhbCppStagingInboundRepository = xhbCppStagingInboundRepository;
     }
 
     protected AbstractControllerBean(EntityManager entityManager) {
@@ -112,5 +119,19 @@ public class AbstractControllerBean {
             xhbFormattingRepository = new XhbFormattingRepository(getEntityManager());
         }
         return xhbFormattingRepository;
+    }
+
+    protected XhbPddaMessageRepository getXhbPddaMessageRepository() {
+        if (xhbPddaMessageRepository == null) {
+            xhbPddaMessageRepository = new XhbPddaMessageRepository(getEntityManager());
+        }
+        return xhbPddaMessageRepository;
+    }
+
+    protected XhbCppStagingInboundRepository getXhbCppStagingInboundRepository() {
+        if (xhbCppStagingInboundRepository == null) {
+            xhbCppStagingInboundRepository = new XhbCppStagingInboundRepository(getEntityManager());
+        }
+        return xhbCppStagingInboundRepository;
     }
 }
