+++ src/main/java/uk/gov/hmcts/pdda/business/services/publicdisplay/AbstractPdConfigReposControllerBean.java
@@ -57,7 +57,7 @@ public abstract class AbstractPdConfigReposControllerBean extends AbstractContro
         PublicDisplayNotifier publicDisplayNotifier, VipDisplayDocumentQuery vipDisplayDocumentQuery,
         VipDisplayCourtRoomQuery vipDisplayCourtRoomQuery) {
         super(entityManager, xhbClobRepository, null, xhbCourtRepository, xhbConfigPropRepository,
-            xhbCppFormattingRepository);
+            xhbCppFormattingRepository, null, null);
         this.xhbRotationSetsRepository = xhbRotationSetsRepository;
         this.xhbRotationSetDdRepository = xhbRotationSetDdRepository;
         this.xhbDisplayTypeRepository = xhbDisplayTypeRepository;
