+++ src/main/webapp/WEB-INF/scheduler.properties
@@ -5,7 +5,7 @@
 #likely to be edited by the application deployer.
 ##
 
-scheduledtasks.pdda=transformtask,cppformattingtask,cppstagingtask,pddabaistask,dlnotifiertask
+scheduledtasks.pdda=transformtask,cppformattingtask,cppstagingtask,pddabaistask,dlnotifiertask,lighthousetask
 
 
 ##
@@ -60,6 +60,17 @@ dlnotifiertask.delay=60000
 dlnotifiertask.period=5000
 dlnotifiertask.hour=6
 dlnotifiertask.minute=30
+
+##
+#Configuration for LighthousePddaControllerBean
+#Description: This task is for inserting CPP data into XHB_CPP_STAGING_INBOUND and
+#updating XHB_PDDA_MESSAGE
+##
+lighthousetask.strategy=uk.gov.hmcts.framework.scheduler.RemoteSessionTaskStrategy
+lighthousetask.remotehome=uk.gov.hmcts.pdda.business.services.pdda.lighthouse.LighthousePddaControllerBean
+lighthousetask.fixedrate=false
+lighthousetask.delay=60000
+lighthousetask.period=30000
 
 
 ############################################################
